requirements.txt
requests>=2.6.0
click>=6.7
beautifulsoup4>=4.6.0
lxml==4.4.1
SQLAlchemy==1.3.10
incremental>=16.10.1
scrapy
scrapy-rotating-proxies==0.6.2
scrapy-useragents
requests_html
